Addressing Open Changes
As we see here to implement a functional hashtable referenced to a
HashTable in Oracle's java.util. Thus all of the elements of table are
stored in static-length array , that s optimally prime.
This example uses a an array for introduction but later , a Vector would
be used to expand logically with scaling.