Address Open Indexing
As we see here to implement a functional hash table to reference the
HashTable built into Oracle's java.util.
Thus all of the elements are store within a static-length array, which
is optimally prime.
This example uses an array for inroduction and efficientcy but later on,
a Vector implementation would be used to expand logically with scaling.