ARC::FunctionLocate
Where we want to avoid secondary clustering, and it as also neccessary
to implement another technique, called double hashing, whereby to use a
second hash function to find the magnitude of a constant offset value.

No easy task to accomplish with arbitrary keys as there is only a single
provided hashCode function. Additionally, multiples and and factors of
the hash table must be avoided to prevent the locate function from
entering into an infinite loop.