ARC::Rehash_loadFactor
As a matter purely of maintaining good performance, the load factor must
be as low as possible. The loadFactor is our constant maximum allowable
rate of load. If this boundary is exceeded the array is then expanded
(method rehash)
To add a value, we first tind the appropriate slot and insert the new
association. If the optimal slot is already occupied (must contain
equivalent key), we return the replaced association. If the reference to
an empty cell is replaced by the reserved association, null is instead
returned.