ARC::getFunction
In order the accurately measure the difficulty of locating an empty slot
using hashing, a function named load factor, zeta, is used to compute
the ratio of the number of values currently in the the table to the
number of slots occupied. Whilst addressing openly, the load factor must
not exceed 1.