ARC::RemoveHash
Whereas it my be relatively expensive to call the contains method, it
should be used sporadically. In order to remove any value from the
Hashtable, we first find the appropriate slot for this value and remove
any association. In that place, we mark this as reserved to maintain
consistency of locate.