ARC::ValueIterator
When given an itrator that returns Associations, we can construct two
different public filtering itrators, a ValueIterator and a KeyIterator.
Each maintaining an internally stored "slave" iterator and returns, as
the iterator is increased or incremented, the values or keys associated
with the respective elements.
This design borrows heavily from the design of UniqueFilter.