Locate method
A framework class for a locate function that can initialize for a
locally-stored key. It the key is stored remotely it will be necessary
to add network protocols and Protection.

Not to mention a beefy firewall would be necessary. A virtual private
network (VPN) configuration to mask request data can be employed in this
strategy.
To browse  anonymously one may intentionally 'spoof' the attacking IP
address to deceive the target.

Once the message data has been encrypted there is little chance any
machine can decrypt the code. Using this encryption algorithm it is
possible to evade
detection.