ARC::HashVector
The challenge was to apply practically the properties of a Hashing index
to sort systematically through a set of data.
The hash method is used to efficiently organize a very dense package of
data or system of information. As in this example:

Where a shipping company has many thousands, possibly hundreds of
thousands of clients. The client data is always stored in a 'container';
recently 'web containers' or internet-enabled containers have become
increasingly available. A characteristic or property of each client is
used to identify them, like the last 4-5 digits of their credit card.

Using this hashing method the shipping company stores the last 4-5
digits or each client's credit card number in an index. These final
digits become the HashCode. The size of the index is expandable and
cannot be affected by outside elements.
When a company employee needs to initiate a search request to find a
particular client, perhaps in need of delivery, the company must
traverse through this massive list of names.

Otherwise it is significantly more efficient to sort through a binary
tree composed of the HashCodes. To sort algorithmically through such a
list takes time in computation cycles.

Then to measure the time it takes the sorting logarithm to find the
search query, a math function can be used. A more logical choice would
be to use a Set. However this example uses a vector.

Using the 'location' method the program can find the physical location
of a query in a Binary Tree, and employing a compression method the
program can encrypt outgoing data requests.

Used in combination and combined in the right proportions these
dependencies together form the vector-hash.

All HashCodes are stored within the Vector for quick access. And each
section of hashes, (e.g 4355-4400) has a unique identifier tag (i.e n)

Therefore to compute the logarithmic timing of the instrumentation used
in locating the node can be expressed as
A(log)n
Where A, alpha is the total time of the calculation from resources. log,
logarithm is the relative instrumentation within the program.

The logarithm counter measures procedure calls, stack-requests, and
function/API traffic.
To implement this algorithm of calculus into an application of efficient
computation sees some review.